Capture output from stdlib `log` library.
Some components, particularly `net/http`, occasionally output log lines
via `log.Print`. We'd like to capture these and send them to rsyslog so
all our log data goes to the same place, and so that we can attach log
line checksums to them.

This uses log.SetOutput to change the log output to a `bytes.Buffer`,
then consumes that buffer line-by-line in a goroutine and sends it to
our rsyslog logger.