Fix Coveralls branch info and update README badges
goveralls was not able to detect the current branch since #4735. Instead of
figuring out which variables are needed for it to work correctly, this commit
just propagates all environment variables with the TRAVIS_ prefix.

Also updates the badges in the readme file. The badge wasn't update when we
migrated to travis-ci.com. The travis-ci.org address now just has an annoying
redirect page which requires you to click a link to go to the new page on
travis-ci.com. The new badge just takes you there directly.