Add -expect-reason flag to checkocsp
Adds a new -expect-reason flag to the checkocsp binary to allow for
verifying the revocation reason of the certificate(s) in question.
This flag has a default value of 0, corresponding to the "Unspecified"
reason.

The behavior to actually verify the revocation reason is gated on the
flag actually being specified on the command line; verifying the
revocation reason is not always desired, so checkocsp needs a way to
differentiate between the flag having its default value explicitly set
versus falling back to its default value.

Fixes #4885