Add test for re-signed OCSP revocation reasons
In https://bugzilla.mozilla.org/show_bug.cgi?id=1648840, it was seen
that Let's Encrypts OCSP responder was generating OCSP responses with
an empty revocationReason field, but only for responses which had been
re-signed (i.e. it was correctly providing a revocationReason for the
first few days after a revocation).

This test issues and then revokes a cert, then jumps 20 days forward
to force the ocsp-updater to re-sign the ocsp response. It then checks
that the new response still has the correct revocation reason.

Fixes #4907