Revert "Update CA RPC interface to proto3 (#4951)"
This reverts commit 24e782e8b4cfe8cb887bc66378916087cc100ea9.

During a rollout, not all boulder components are updated
simultaneously. If an updated proto3 ra sends an IssueCertificateRequest
with OrderID 0 (always the case for ACMEv1 requests, because Orders
are an ACMEv2 concept), but a non-updated proto2 ca receives that
request, it will fail the non-nil check in the wrapper function and
result in an IncompleteRequest error.

After this is reverted, we will proceed with proto3 migrations by
first removing non-nil checks from the wrapper functions, then updating
the protos themselves after that change has been deployed.

Concerns #4940