Remove non-nil checks from proto wrappers
We are intending to migrate all boulder services from proto2 to
proto3. One of the advantages of proto3 is that the generated Go
code doesn't use indirection for built-in types: e.g. it uses int64
instead of *int64 to represent a proto's int64 field. However, this
means that many fields are no longer nillable, and will instead take
their default zero-value when left unset. Some of these zero-values
are valid values, and therefore should not result in the server
complaining with `errIncompleteRequest`.

This change simplifies the completeness checks around in-memory
protos. This should make it possible for a proto2 and a proto3
service to cross-communicate without resulting in errors.

Issues: #4940, #4955, #4956