Enable exportloopref linter
The golangci-lint metapackage includes exportloopref at v0.1.17,
the latest tagged release of exportloopref. Enabling it in our
config should be all that is necessary to start benefitting from
it.

At this time, exportloopref does not detect cases where a reference
to a loop var is passed to a function. I'm investigating contributing
that support upstream (and then bumping the version bundled with
golangci-lint).

Fixes #4948