Anchor all referenced loop variables
Taking a reference to a loop variable produced by a range statement
can be dangerous. If that pointer is preserved elsewhere, for example
by being passed to a function which stores it, the resulting behavior
may not be what the author or reader expects.

When it is not feasible to structure code such that references to
loop variables are not taken at all, the loop variable can instead be
'anchored' by assigning it to itself.

This change anchors all loop range variables which are referenced and
then passed to arbitrary functions. It also removes our exclusion of
gosec'c G601 lint rule, which catches this error, to ensure we don't
reintroduce it later.

Fixes #4986