sa: remove nil checks for zero-able fields.
As part of the migration to proto3, any fields in requests that may be
zero should also be allowed to be nil. That's because proto3 will
represent those fields as absent when they have their zero value.

This is based on a manul review of the wrappers for the SA.